naive hand-rolled dct hash implementation
The hashes it produces are not compatible with pHash; however,
the hamming distances for several images I've tried match that
of pHash, so it seems fine.

I haven't measured performance, but I expect it to be worse
than pHash. This implementation doesn't precompute the DCT matrix,
for instance.

pHash cannot be used for several reasons:
* it's GPL
* it relies on glibc features (in other words, can't be linked against musl)
* it's been unsupported since 2013 and already requires hacks to keep it
compiling on modern toolchains

image_hash doesn't cut it either:
* using it means maintaining a suboptimal wrapper that can't
even return Result where it's needed
* for some reason, it compares frequencies against the mean value
(other implementations use the median)
* as a consequence, it reports different images as having closer
hamming distance than other impls do

I still need to give this a spin on a larger set of images.